requirements.txt
pandas~=2.3.3
pyautogui~=0.9.54
pyperclip~=1.11.0
pynput~=1.8.1
PyYAML~=6.0.3
